Update project ver 1.0

+++ WebBanHang/WebBanHang/Areas/Admin/Controllers/BrandController.cs
@@ -42,7 +42,7 @@ namespace WebBanHang.Areas.Admin.Controllers
             return View();
         }
 
-        public async Task<IActionResult> Edit(int Id)
+        public async Task<IActionResult> Edit(long Id)
         {
             BrandModel brand = await _dataContext.Brands.FindAsync(Id);
             return View(brand);
@@ -142,7 +142,7 @@ namespace WebBanHang.Areas.Admin.Controllers
         #endregion
 
         #region Delete
-        public async Task<IActionResult> Delete(int Id)
+        public async Task<IActionResult> Delete(long Id)
         {
             BrandModel brand = await _dataContext.Brands.FindAsync(Id);
             _dataContext.Brands.Remove(brand);
