+++ WebBanHang/WebBanHang/Areas/Admin/Controllers/ProductController.cs
@@ -103,7 +103,7 @@ namespace WebBanHang.Areas.Admin.Controllers
             }
         }
 
-        public async Task<IActionResult> Edit(int Id)
+        public async Task<IActionResult> Edit(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             ViewBag.Categories = new SelectList(_dataContext.Categories, "Id", "Name", product.CategoryId);
@@ -202,7 +202,7 @@ namespace WebBanHang.Areas.Admin.Controllers
             }
         }
 
-        public async Task<IActionResult> Delete(int Id)
+        public async Task<IActionResult> Delete(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
