+++ WebBanHang/WebBanHang/Controllers/CartController.cs
@@ -23,7 +23,7 @@ namespace WebBanHang.Controllers
             return View(cartVM);
         }
 
-        public async Task<IActionResult> Add(int Id)
+        public async Task<IActionResult> Add(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart") ?? new List<CartItemModel>();
